Update README.md
This commit adds a description and link to the already existing
document that details the installation of gitlab on ICP, namely
docs/deploy-with-ICP.md

User paulczar initially submitted this branch in PR#48, but the
pull request was closed before it could be merged.  This commit
merely reintroduces this branch with updates from master since its
initial submission.